clean up Compliance in as_regex
